+++ packages/openai/src/openai-provider.ts
@@ -3,8 +3,8 @@ import {
   ImageModelV3,
   LanguageModelV3,
   ProviderV3,
-  SpeechModelV2,
-  TranscriptionModelV2,
+  SpeechModelV3,
+  TranscriptionModelV3,
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
@@ -81,12 +81,12 @@ Creates a model for image generation.
   /**
 Creates a model for transcription.
    */
-  transcription(modelId: OpenAITranscriptionModelId): TranscriptionModelV2;
+  transcription(modelId: OpenAITranscriptionModelId): TranscriptionModelV3;
 
   /**
 Creates a model for speech generation.
    */
-  speech(modelId: OpenAISpeechModelId): SpeechModelV2;
+  speech(modelId: OpenAISpeechModelId): SpeechModelV3;
 
   /**
 OpenAI-specific tools.