+++ packages/perplexity/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/perplexity
 
+## 2.1.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [046aa3b]
+- Updated dependencies [e8109d3]
+  - @ai-sdk/provider@2.1.0-beta.5
+  - @ai-sdk/provider-utils@3.1.0-beta.7
+
 ## 2.1.0-beta.6
 
 ### Patch Changes